fix x11 race condition
This function would previously try to avoid race conditions by checking if `connection != NULL`, then locking the mutex, then checking the same condition again. The design seemed okay at first, but in an edge-case it could result in one thread writing to `connection` while another is reading it to check if it's NULL, thus invoking undefined behaviour.

Locking a mutex on every native call sounds like it'd be a bad thing, but it actually has no performance cost as long as there's no lock contention - and we usually only call native functions from the main thread.